Fix eager loading issue ancestors/descendants relationships (fixes #339, #351, #329, #334, #271)
The way eager loading is handled in Laravel has changed since the
original implementation of the NestedSet library. It was not possible
anymore to apply the scope to eagerly loaded ancestors/descendants
relationships.

Also removed defaultOrder() from the relations so custom orders can be
applied. Sorting should never be a part of a relation anyway, or it
would be hard to customize.